chore(...) : 청소하기-signUp part 정리 (재플린 참조, adjustResize 고려 layout 변경) , 툴바적용. clear 색상 변경
