+++ app/src/main/res/layout/activity_sign_up.xml
@@ -13,13 +13,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/appBar"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:translationZ="90dp"
-            app:layout_constraintTop_toTopOf="parent" />
-
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/container_sign_up"
             android:layout_width="0dp"
